Verify King's move is valid
Per the rules of classical chess, a king can not make a move that would cause it to be attacked.  I've added logic to discard such moves and have added a few test cases to validate the behavior.
NOTE: This is definitely not the best code I've ever written, but it does the job.